+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/admin-frontend/compare/0.1.7...develop)
+## [Unreleased](https://github.com/ixocreate/admin-frontend/compare/0.1.8...develop)
+
+## [v0.1.8 (2019-07-22)](https://github.com/ixocreate/admin-frontend/compare/0.1.7...0.1.8)
+### Changed
+- Media `size` -> `fileSize`
+### Added
+- MultiSelect createNewDeferred option
 
 ## [v0.1.7 (2019-07-11)](https://github.com/ixocreate/admin-frontend/compare/0.1.6...0.1.7)
 ### Changed